requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
pydantic>=2.0.0
python-dateutil>=2.8.2
